+++ lib/page-acala/acalaEntry.dart
@@ -43,7 +43,16 @@ class AcalaEntry extends StatelessWidget {
               child: Observer(
                 builder: (_) {
                   if (store.settings.loading) {
-                    return CupertinoActivityIndicator();
+                    return Container(
+                      padding: EdgeInsets.only(
+                          top: MediaQuery.of(context).size.width / 2),
+                      child: Column(
+                        children: [
+                          CupertinoActivityIndicator(),
+                          Text(I18n.of(context).assets['node.connecting']),
+                        ],
+                      ),
+                    );
                   }
                   return ListView(
                     padding: EdgeInsets.all(16),